requirements.txt
aiodns==3.1.1
aiohttp==3.9.0
aiopg==1.4.0
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
ccxt==4.0.80
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.1.0
colorlog==6.8.0
cryptography==41.0.7
discord.py==2.3.2
discord.py-pagination==0.0.5
disnake==2.9.1
frozenlist==1.3.3
greenlet==3.0.3
idna==3.4
joblib==1.3.2
kaleido==0.2.1
multidict==6.0.4
numpy==1.26.3
packaging==23.2
plotly==5.16.1
psycopg2==2.9.9
psycopg2-binary==2.9.7
pycares==4.4.0
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
SQLAlchemy==2.0.25
ta==0.11.0
tenacity==8.2.3
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
yarl==1.9.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
seaborn==0.13.1
patsy==0.5.6
statsmodels==0.14.1
